requirements.txt
streamlit
tensorflow
numpy
pillow
sqlite3
pandas